Rename designTokens directory to tokens
