--- src/designTokens/index.ts
@@ -1,5 +1,0 @@
-export * from './colorSchemes';
-
-export * from './designTokens';
-
-export * from './motion';
