+++ src/tokens/_typography.ts
@@ -1,6 +1,6 @@
-import type {Tokens} from './designTokens';
+import type {TokenObject} from './tokens';
 
-export const typography: Tokens = {
+export const typography: TokenObject = {
   'font-weight-regular': '400',
   'font-weight-medium': '500',
   'font-weight-semibold': '600',
